Updated dependencies, Hopefully fixed a bug where adapter did not exit as expected
