+++ package.json
@@ -19,37 +19,37 @@
     "url": "https://github.com/crycode-de/ioBroker.odl"
   },
   "dependencies": {
-    "@iobroker/adapter-core": "^2.1.0",
+    "@iobroker/adapter-core": "^2.4.0",
     "core-decorators": "^0.20.0",
-    "request": "^2.88.0"
+    "request": "^2.88.2"
   },
   "devDependencies": {
-    "@iobroker/testing": "^2.0.1",
-    "@types/chai": "^4.2.7",
-    "@types/chai-as-promised": "^7.1.0",
-    "@types/gulp": "^4.0.6",
-    "@types/mocha": "^5.2.7",
-    "@types/node": "^10.17.13",
+    "@iobroker/testing": "^2.3.0",
+    "@types/chai": "^4.2.12",
+    "@types/chai-as-promised": "^7.1.3",
+    "@types/gulp": "^4.0.7",
+    "@types/mocha": "^8.0.3",
+    "@types/node": "^10.17.35",
     "@types/proxyquire": "^1.3.28",
-    "@types/request": "^2.48.4",
-    "@types/sinon": "^7.5.1",
-    "@types/sinon-chai": "^3.2.2",
-    "@typescript-eslint/eslint-plugin": "^2.17.0",
-    "@typescript-eslint/parser": "^2.17.0",
-    "axios": "^0.19.2",
+    "@types/request": "^2.48.5",
+    "@types/sinon": "^9.0.7",
+    "@types/sinon-chai": "^3.2.5",
+    "@typescript-eslint/eslint-plugin": "^4.3.0",
+    "@typescript-eslint/parser": "^4.3.0",
+    "axios": "^0.20.0",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
-    "eslint": "^6.8.0",
-    "eslint-plugin-html": "^6.0.0",
+    "eslint": "^7.10.0",
+    "eslint-plugin-html": "^6.1.0",
     "gulp": "^4.0.2",
-    "mocha": "^7.0.0",
+    "mocha": "^8.1.3",
     "proxyquire": "^2.1.1",
-    "rimraf": "^3.0.0",
-    "sinon": "^8.1.1",
-    "sinon-chai": "^3.4.0",
-    "source-map-support": "^0.5.16",
-    "ts-node": "^8.6.2",
-    "typescript": "^3.7.5"
+    "rimraf": "^3.0.2",
+    "sinon": "^9.1.0",
+    "sinon-chai": "^3.5.0",
+    "source-map-support": "^0.5.19",
+    "ts-node": "^9.0.0",
+    "typescript": "^4.0.3"
   },
   "main": "build/main.js",
   "scripts": {
